+++ packages/cli/src/logger.js
@@ -8,7 +8,11 @@ const loggerOptions = {
     default: 'info',
     type: 'string',
     choices: ['trace', 'debug', 'info', 'warn', 'error', 'silent'],
-    describe: 'console log level'
+    describe: 'Console log level'
+  },
+  'log-json-format': {
+    boolean: true,
+    describe: 'Log output format in json'
   },
   'log-file': {
     alias: 'L',
@@ -22,39 +26,33 @@ const loggerOptions = {
   }
 }
 
-const addFileLogger = async (file, level) => new Promise(resolve => logger.addFile(file, level, resolve))
+const addFileLogger = async (file, level = 'info') => new Promise(resolve => logger.addFile(file, level, resolve))
 
 const loggerMiddleware = async (argv) => {
   const logMessages = []
 
-  const enabledLoggers = {
-    console: false,
-    file: false
-  }
-
-  if (argv.logLevel) {
+  if (argv.logLevel && argv.logJsonFormat) {
+    logger.addJson(argv.logLevel)
+    logMessages.push(`Add json console logger in level ${argv.logLevel} through cli args`)
+  } else if (argv.logLevel) {
     logger.addPretty(argv.logLevel)
     logMessages.push(`Add console logger in level ${argv.logLevel} through cli args`)
-    enabledLoggers.console = true
   }
   if (argv.logFile) {
     await addFileLogger(argv.logFile, argv.logFileLevel)
-    logMessages.push(`Add file logger in level ${argv.logLevel} and file ${argv.logFile} through cli args`)
-    enabledLoggers.file = true
+    logMessages.push(`Add file logger in level ${argv.logFileLevel || 'info'} and file ${argv.logFile} through cli args`)
   }
 
-  if (argv.config) {
-    const { config } = load(argv.config).catch(()=> ({}))
+  if (argv.config && !logMessages) {
+    const { config } = await load(argv.config).catch(()=> ({}))
     const loggers = config?.logger || []
     for (const configLogger of loggers) {
-      if (enabledLoggers[configLogger.type]) {
-        logMessages.push(`Skip ${configLogger.type} logger by config. Logger was already initialized through cli args`)
-        continue
-      }
-
-      if (configLogger.type == 'console') {
+      if (configLogger.type == 'console' && configLogger.format == 'json') {
+        logger.addJson(configLogger.level)
+        logMessages.push(`Add json console logger in level ${configLogger.level} through config`)
+      } else if (configLogger.type == 'console') {
         logger.addPretty(configLogger.level)
-        logMessages.push(`Add console logger in level ${configLogger.level} through cli args`)
+        logMessages.push(`Add console logger in level ${configLogger.level} through config`)
       } else if (configLogger.type == 'file') {
         await addFileLogger(configLogger.file, configLogger.level)
         logMessages.push(`Add file logger in level ${configLogger.level} and file ${configLogger.file} through config`)