+++ packages/cli/src/tasks/server.js
@@ -1,13 +1,22 @@
-const { runCli, loggerArgs } = require('./run')
+const { spawnCli } = require('./run')
+
+const getCommonCliArgs = options => {
+  const { configFile, autoConfigFile } = options
+  const configArgs = !autoConfigFile ? ['-c', configFile] : []
+
+  return [...configArgs]
+}
 
 const startServer = async options => {
-  const { config, configFile, autoConfigFile } = options
-  const args = [...loggerArgs(config), 'server']
+  const commonArgs = getCommonCliArgs(options)
 
-  if (!autoConfigFile) {
-    args.push('-c', configFile)
-  }
-  await runCli(args)
+  await new Promise((resolve, reject) => {
+    const serverProcess = spawnCli([...commonArgs, 'server'])
+
+    serverProcess.once('exit', (code) => {
+      code == 0 ? resolve() : reject(new Error('Server exited with code ' + code))
+    })
+  })
   return 'exit'
 }