+++ packages/logger/src/json-stream.js
@@ -1,0 +1,5 @@
+const createJsonStream = (rootLogger, level) => {
+  rootLogger.add({level, stream: process.stdout})
+}
+
+module.exports = createJsonStream